streamlit_app/requirements.txt
streamlit==1.31.0
clickhouse-connect==0.7.0
kafka-python==2.0.2
pandas==2.2.0
plotly==5.18.0
pytz==2024.1
pymongo==4.6.1
scikit-learn==1.4.0
joblib==1.3.2
numpy==1.26.3